Basic separation of existing functionality into the two apps
