+++ folk_rnn_site/archiver/__init__.py
@@ -1,0 +1,1 @@
+MAX_RECENT_ITEMS = 5
