+++ folk_rnn_site/composer/forms.py
@@ -35,21 +35,3 @@ class ComposeForm(forms.Form):
             tokens = self.cleaned_data['prime_tokens'].split(' ')
             if not validate_tokens(tokens, model_file_name=self.cleaned_data['model']):
                 self.add_error('prime_tokens', ValidationError('Invalid ABC as per RNN model', code='invalid'))
-
-class TuneForm(forms.Form):
-    tune = forms.CharField(label='', widget=forms.Textarea(attrs={'id': 'abc'}))
-    edit = forms.ChoiceField(label='', 
-                            widget=forms.RadioSelect(attrs={'onchange': 'this.form.submit();'}), 
-                            choices=(
-                                ('rnn', 'RNN original'),
-                                ('user', 'Your setting'),
-                                ))
-    edit_state = forms.ChoiceField(widget=forms.HiddenInput,
-                            choices=(
-                                ('rnn', ''),
-                                ('user', ''),
-                                ))
-    
-class CommentForm(forms.Form):
-    text = forms.CharField(label='Comment:', widget=forms.Textarea(attrs={'id': 'new_comment'}))
-    author = forms.CharField(label='Author:', widget=forms.TextInput(attrs={'id': 'new_comment_author'}))
