+++ folk_rnn_site/composer/templates/composer/_abcjs-bodyend.html
@@ -1,0 +1,30 @@
+        <script type="text/javascript">
+            function selectionCallback(abcelem) {
+                var note = {};
+                for (var key in abcelem) {
+                    if (abcelem.hasOwnProperty(key) && key !== "abselem")
+                        note[key] = abcelem[key];
+                }
+                console.log(abcelem);
+                var el = document.getElementById("selection");
+                el.innerHTML = "<b>selectionCallback parameter:</b><br>" + JSON.stringify(note);
+            }
+        
+            function initEditor() {
+                new ABCJS.Editor("abc", { paper_id: "paper0",
+                    generate_midi: true,
+                    midi_id:"midi",
+                    midi_download_id: "midi-download",
+                    generate_warnings: true,
+                    warnings_id:"warnings",
+                    midi_options: {
+                        generateDownload: true
+                    },
+                    render_options: {
+                        listener: { highlight: selectionCallback }
+                    }
+                });
+            }
+        
+            window.addEventListener("load", initEditor, false);
+        </script>
