+++ folk_rnn_site/composer/templates/composer/_abcjs-headend.html
@@ -1,0 +1,9 @@
+        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.5.0/css/font-awesome.min.css">
+        <link href="/static/abcjs-midi.css" media="all" rel="stylesheet" type="text/css" />
+        <script src="/static/abcjs_editor_midi_3.1.4-min.js" type="text/javascript"></script>
+        <script type="text/javascript">window.ABCJS.midi.soundfontUrl = "/static/soundfont/"</script>
+        <style>
+            .abcjs-inline-midi {
+                max-width: 740px;
+            }
+        </style>
