+++ folk_rnn_site/composer/templates/composer/home.html
@@ -1,0 +1,9 @@
+{% extends "composer/base.html" %}
+
+{% block content %}
+{% load humanize %}
+<div id="compose">
+    <h1>Compose</h1>
+    {% include "composer/_compose.html" %}
+</div>
+{% endblock %}
