+++ folk_rnn_site/composer/templates/composer/tune-in-process.html
@@ -1,4 +1,4 @@
-{% extends "base.html" %}
+{% extends "composer/base.html" %}
 
 {% block title %}Candidate Tune{% endblock %}
 {% block headend %}<meta http-equiv="refresh" content="3">{% endblock %}
