+++ folk_rnn_site/composer/templates/composer/tune.html
@@ -1,0 +1,45 @@
+{% extends "composer/base.html" %}
+
+{% block title %}{{ tune.title }} on Folk RNN{% endblock %}
+
+{% block headend %}{% include "composer/_abcjs-headend.html" %}{% endblock %}
+{% block bodyend %}{% include "composer/_abcjs-bodyend.html" %}{% endblock %}
+
+{% block content %}
+{% load humanize %}
+<h1>{{ tune.title }}</h1>
+<div id="compose_ui">
+    <div id="composition">
+        <p>The composition –</p>
+        <pre id="abc">{{ tune.abc }}</pre>
+    </div>
+    <div id="composition_metadata">
+        <p>Was made with –<br>
+            <ul>
+                <li>RNN model: {{ tune.rnn_model_name }}</li>
+                <li>RNN seed: {{ tune.seed }}</li>
+                <li>RNN temperature: {{ tune.temp }}</li>
+                <li>Prime tokens: {{ tune.prime_tokens }}</li>
+            </ul>
+        </p>
+        <p>The task was –<br>
+            <ul>
+                <li>Requested at: {{ tune.requested|date:"r" }}</li>
+                <li>Composition took: {{ rnn_duration|floatformat:"0" }}s</li>
+            </ul>
+        </p>
+    </div>
+    <div id="midi"></div>
+    <div id="midi-download"></div>
+    <div id="warnings"></div>
+    <div id="music"></div>
+    <div id="paper0"></div>
+    <div id="paper1"></div>
+    <div id="paper2"></div>
+    <div id="paper3"></div>
+    <div id="selection"></div>
+    <form action="/">
+       <input id="compose_button" type="submit" value="Start again...">
+    </form> 
+</div>
+{% endblock %}
