--- folk_rnn_site/composer/templates/tune.html
@@ -1,82 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}{{ tune.title }} on Folk RNN{% endblock %}
-
-{% block headend %}{% include "_abcjs-headend.html" %}{% endblock %}
-{% block bodyend %}{% include "_abcjs-bodyend.html" %}{% endblock %}
-
-{% block content %}
-{% load humanize %}
-<h1>{{ tune.title }}</h1>
-<div id="compose_ui">
-    <div id="composition">
-        <p>The composition –</p>
-        <form method="POST">
-            {% csrf_token %}
-            {{ tune_form }}
-            {% ifequal tune_form.edit.data 'user' %}
-                <input id="save_button" name="save" type="submit" value="Save">
-                <input id="setting_button" type="submit" name="submit_setting" value="Submit Setting">
-            {% endifequal %}
-        </form>
-    </div>
-    <div id="composition_metadata">
-        <p>Was made with –<br>
-            <ul>
-                <li>RNN model: {{ tune.rnn_model_name }}</li>
-                <li>RNN seed: {{ tune.seed }}</li>
-                <li>RNN temperature: {{ tune.temp }}</li>
-                <li>Prime tokens: {{ tune.prime_tokens }}</li>
-            </ul>
-        </p>
-        <p>The task was –<br>
-            <ul>
-                <li>Requested at: {{ tune.requested|date:"r" }}</li>
-                <li>Composition took: {{ rnn_duration|floatformat:"0" }}s</li>
-            </ul>
-        </p>
-    </div>
-    <div id="midi"></div>
-    <div id="midi-download"></div>
-    <div id="warnings"></div>
-    <div id="music"></div>
-    <div id="paper0"></div>
-    <div id="paper1"></div>
-    <div id="paper2"></div>
-    <div id="paper3"></div>
-    <div id="selection"></div>
-    <form action="/">
-       <input id="compose_button" type="submit" value="Start again...">
-    </form> 
-</div>
-
-<div id="settings">
-{% if settings %}
-    <h2>{{ settings|length|apnumber|capfirst }} setting{{ settings|length|pluralize }}</h2>
-    {% for setting in settings %}
-        <div id="setting {{ setting.header_x }}">
-            <pre>{{ setting.abc }}</pre>
-            <!-- TODO: Added by..., Download..., Sheet music..., -->
-        </div>
-    {%  endfor %}
-{% else %}
-    <h2>No settings</h2>
-{%  endif %}
-</div>
-
-<div id="comments">
-    <h2>Comments</h2>
-    <form method="POST">
-        {% csrf_token %}
-        {{ comment_form }}
-        <input id="comment_button" type="submit" name="submit_comment" value="Submit">
-    </form>
-    {% if comments %}
-    <ul id="comment_list">
-        {% for comment in comments %}
-        <li>{{ comment.text }}<br>— {{ comment.author }}, {{ comment.submitted|naturalday }}</li>
-        {% endfor %}
-    </ul>
-    {% endif %}
-</div>
-{% endblock %}
